Major optimization #1
Remove the SafeCheck proxy and be more "flat."

The logic behind this change was that I saw a lot of garbage collection
in the profile and a lot of work being spent in visitJSON even though
we don't have that many JSON checks.

Turns out that when we `return {}` in JSONSyntaxError, we'd visit the
entire tree anyway. The methods were all minor, but they were all pushed
on the event loop since they were all resolving as an "empty" promise.

Also, given that there was a LOT of garbage collection going on, I
decided to be a bit more memory conscious with the visitJSON/visitLiquid
methods and reimplement the depth first traversal by using a stack
instead of recursion. I also was a bit more conscious of the objects I
was needlessly creating (Object.entries calls, array.filter, etc.).

The function is a whole lot more dumb but runs much faster :D