Add LinkedEditingRanges support for HTML tag names (#486)
Just like HTML, this feature remains behind the `editor.linkedEditing` configuration.

```json
// In VS Code's settings.json
"editor.linkedEditing": true
```

Reason being is that VS Code's implementation is kind of buggy. The
thing can get out of sync or you can type the name of a void element and
all of a sudden the rename doesn't complete anymore.

This stuff is hard to explain and I suspect is the reason why MSFT chose
to put that setting opt-in rather than opt-out.