Fix HTML element close tag auto-insertion inside Liquid branches
There was a bug with our findCurrentNode algorithm that made it so we'd go in the wrong node when a branch had child nodes.

The fix was to have more information on the AST and flag LiquidBranch as "unclosed" when their blockEndPosition is -1, -1.

It's kind of a nasty bug tbh.

Fixes #515