Remove getDefault(Schema)?Translations dependencies
We still have an optiomal dependency that is otherwise implemented by a
naive implementation (for CLI use cases), but in the language server we
prefer the buffer over the file in the file system...

Which makes me think we could get rid of that by actually making that
logic part of the getTranslationsFactory directly. Prefer theme files
over fs.