Virtual File Systems & Shopify Liquid VS Code extension for Web (#529)
* Simplify theme-language-server-browser connection creation

* Add VS Code for Web Extension skeleton implementation

* Remove support for ruby-based language server

* Add VS Code for Web major bump changeset

* Fixup the sourcemaps & launch.json for dual build

* Clean up asset-size-based tests

* Include theme check test helpers in dist

* Replace `fileExists` dependency by `fs` injection

- Change `Config` interface to refer to `rootUri` instead of `root` as an absolutePath
- Get rid of `fileExists` DI in theme-check Dependencies
- Get rid of `fileExists` DI in theme-language-server Dependencies
- Add `fs` DI in theme-check Dependencies
- Add `fs` DI in theme-language-server Dependencies
- Add `MockFileSystem` implementation (for tests)
- Add `NodeFileSystem` implementation
- Add optional `fs` argument to `theme-language-server-node#startServer()`

The `fs` DI has the following interface (so far):

```ts
interface FileSystem {
  stat(uri: string): Promise<{ fileType, size }>;
  readFile(uri: string): Promise<string>;
  readDirectory(uri: string): Promise<[uri: string, fileType: FileType][]>;
}
```

theme-language-server-node now also accepts an optional `fs` implementation in startServer.

* Get rid of the fileSize injection (use abstract `fs.stat` instead)

* Replace getDefault(Schema)?Locale dependency injection

Instead of having that logic be injected in, I added an abstract
implementation in theme-check-common that depends on the virtual file
system. We will no longer need to inject this behaviour :)

* Rename FileSystem to AbstractFileSystem

FileSystem is an ambient import which made the DX suffer when
refactoring code. It's also more obvious now that the FileSystem
interface is an interface that you need to implement, not something
concrete that you can use directly.

* Replace AbsolutePath concern in theme-check-common

Replace it with URIs. That way we won't need to worry about windows paths being weird.

The only place where we still have them is for places where we deal with `glob` and `loadConfig` in `theme-{check,language-server}-node`.

- SourceCode.absolutePath -> SourceCode.uri
- etc.

* Fixup baseUrl in tcn

* Normalize URIs for windows and remove back slashes of relative paths

* Uri->UriString for clarity

* Remove getDefault(Schema)?Translations dependencies

We still have an optiomal dependency that is otherwise implemented by a
naive implementation (for CLI use cases), but in the language server we
prefer the buffer over the file in the file system...

Which makes me think we could get rid of that by actually making that
logic part of the getTranslationsFactory directly. Prefer theme files
over fs.

* Move buffer translations concern to theme-check-common

All that logic didn't need to be in the language server.

This makes the dep injection easier too because we no longer have this
upstream concern.

* Remove filesForURI language server dependency

Replace with AbstractFileSystem-based implementation

* Remove getThemeSettingsSchemaForURI injection

Replace with abstract-file-system-based implementation

TODO make that shit faster. Seems highlighy unoptimized.

* Remove findRootURI injection

Replace with abstract-file-system-based implementation in common code.

* Implement and setup basic VsCodeFileSystem

- Adds support for VS Code for Web
- Adds support for remote files

* Optimize FS calls in the Language Server

You don't need to query VS Code _every time_ the user types something
for the same imformation. That's rather slow!

We'll invalidate the relevant caches when files are
saved/deleted/created/renamed.

* Implement MainThreadFileSystem in CodeMirror Playground

Made to demo what it's like to use a VirtualFileSystem in a
non-vscode environment.

* Add a bunch of changesets

* Docs changes and webpack config cleanup

* Add toplevel `yarn dev:web` command

* Minor cleanups

* Add Debug Web Extension in Chromium launch config

* Add support for prettier formatting in web/remote contexts

I also removed the double reporting of parser errors. Now that theme check
is on the same parser as the prettier plugin, there's no point in showing
the error message reported from two different sources.

Fixes #533

* PR Cleanups

* All dev servers on :3000