Fix `loadConfig` for remote files (#543)
We'll steal a page from the prettier playbook and
simply won't load configs in remote settings. The
problem with remote plugins and settings is that
we simply can't run arbitraty node code in browser
or from a `github:` VFS.

What we do is simply ignore remote configs. We'll
offer the recommended checks for remote files. Just
like prettier.

Fixes #532