Do not activate for non-theme-{json,liquid} files
I wrote this thing 3 or 4 years ago without really understanding what I
was doing. This is what the client uses to tell if it should send
requests/notifications to the language server for a given file.

We were sending messages for files that were not supported by us.

`package.json` and so on were sent to the language server and we weren't
even sure they were theme files.

This commit fixes that

Fixes #576