Improve the JSON completion & hover building blocks (#622)
The json-languageservice API was a bit awkward to use. Had a bunch of
weird type problems that nobody should really have to care about.

In this PR, I use an adapter to make folks write Completion & Hover
providers in a manner similar to what we do with liquid files.

It's not 100% the same, but it's similar enough and abstracts the weird
quirks like "return undefined not Promise<undefined> when you can't
hover something".