Update filter completion tests to assert output
We were previously asserting only against the fact that the text edits
were what we expected but we can have a lot more confidence if we also
test that the application of the text edit is what we expect.

This commit hooks into TextDocument and lets us use `applyEdits` to test
that the modified string looks as we expect it to.