TODO -> Write a new commit message for this. - this now includes defautl values as well - we can explain abit more about the rules too
Parsing support for optional param delimiters `[]`
----
- Added a parameter `required` to the LiquidDocParamNode
- Parameters with `[]` around the name will return `required: true`
- Parameters with incomplete delimiters `e.g. ([missingTail)` will map to a `TextNode`

[LiquidDoc] Support default values for optional param names

This is the direction we should go in.

I originally tried accessing everything from the top level in the param node.
This is still the goal, but I tahink this should be done at stage 2 so that we can leverage the recursive nature of how to-AST works.
My original approach attempted to Write everything to the top level in stage 1, but we should be doing this in stage 2.

OTher todos
- See if the grammar can be refactored or cleaned up.
- Remove any outdated mappings.