Add LiquidDoc parsing support for @description annotation -----
We can parse multiple descriptions for a given LiquidDoc, though we will only use the first one if multiple are present.
This logic will come in future PRs.
We are NOT adding support for implicit descriptions, which are descriptions that are provided before any other annotations are provided - reference: https://jsdoc.app/tags-description