Fix bug where theme check would fail for theme app extensions
This bug only surfaces when a developer would manually run the theme
check against a theme app extension from the CLI. Inside the VSCode
extension we were already doing the right thing so you wouldn't notice
it there and the Shopify CLI hardcodes the check when you build your app
so you wouldn't notice it there either.

This adds some logic to where we load the config to check if we're
inside of an app by looking for two app-like config files. If so, we use
`theme-check:theme-app-extension` as our default checks rather than the
recommended. This ensures that our context is appropriately set as 'app'
instead of 'theme' and removes false positives when running the check.