Fix potential race condition
Dreamed of a weird issue where two modules were being traversed in
parallel, say template A and template B.

Consider the following:
- A finds custom-section.liquid
- B finds custom-section.liquid

Since custom-section.liquid is not traversed yet, and because we were
checking graph.modules as our "module cache", both A & B get new
theme module instances, both with references: [], and dependencies: [],

When we call traverse on it after, we might be binding references and
dependencies on the wrong object.

The problem was that I was overloading visiting and module cache into
one thing and I shouldn't have.

This commit fixes that by using a WeakMap to hold references to module
caches by themeGraph reference. When the themeGraph gets GC'ed, the
module cache will get GC'ed.

Fixes the race condition.
instances