Introduce `@shopify/theme-graph` (#956)
* Introduce `@shopify/theme-graph` foundation

A Theme Graph represents the web of dependencies and references present in a Shopify theme.

Dependency: when a module links to other modules in its own text, we call those dependencies. Examples:
- A template file mentioning sections and blocks. That template is dependent on all those sections and blocks.
- A liquid file uses `{% render 'foo' %}`, that liquid file is dependent on `snippets/foo.liquid`

References: It’s the same as a dependency, but the link goes in the other direction.
- A block is referenced by some template or section group
- A snippet is referenced by some other liquid file because that other liquid file uses `{% render 'nameofthisfile' %}`
- A layout is referenced by some template (either implicitly because layout = theme.liquid by default, or explicitly when the property is in the template)

* Add bin/theme-graph script

* Add theme-graph playground

* Replace @swc/core with acorn/acorn-walk

swc/core was complicated to bundle into the vscode builds.

Simplifies visiting too! Big win.

* VS Code integration of Theme Graph

* Fixup NodeFileSystem.readDirectory on Windows

* Add `themeGraph/deadCode` request handler to startServer

* Make it resilient to dead links

`{% render 'unknown' %}` should show a dependency on
"snippets/unknown.liquid" and we should know in the graph that this file
doesn't exist. We can put a warning icon in the dependency tree for
those.

* Refactor theme-graph to allow for `updateGraph`

* Implement updateThemeGraph

Add create/change unit tests.

TODO delete/rename

* Bench preload vs build graph

* Fix acceptsLocalBlocks throwing when section does not exist

* Get rid of update code given buildThemeGraph runs in ~50ms

Micro optimizations were micro and not bug free.

* Clean ups

* Traverse section and blocks presets (+ section.default)

Presets and default also creates explicit dependencies on other blocks

* Add browser support + loading state + lazy preload

* Clean up double boolean into 1 field (preset/indirect ref)

* Add UI support for dead links (links to file that don't exist)

We'll show a warning icon + color

* Add Theme Graph README

* Add how-it-works.md to theme-graph/docs

* Add changesets

* Simplify buildThemeGraph's sections code

* Fixup makeDeadCode

* List dead code dependencies anyway

* Main readme update

* Remove theme graph playground

* Add loadConfig memoization

* Clean up DeadCode request

* Fix potential race condition

Dreamed of a weird issue where two modules were being traversed in
parallel, say template A and template B.

Consider the following:
- A finds custom-section.liquid
- B finds custom-section.liquid

Since custom-section.liquid is not traversed yet, and because we were
checking graph.modules as our "module cache", both A & B get new
theme module instances, both with references: [], and dependencies: [],

When we call traverse on it after, we might be binding references and
dependencies on the wrong object.

The problem was that I was overloading visiting and module cache into
one thing and I shouldn't have.

This commit fixes that by using a WeakMap to hold references to module
caches by themeGraph reference. When the themeGraph gets GC'ed, the
module cache will get GC'ed.

Fixes the race condition.
instances