Fix false reporting of UndefinedObject within visible_if
We were visiting those nodes as though they were part of the execution
of the file when in fact those are not parsed by the backend but
stripped away.

I first tried doing that at the parsing layer by putting the body as a
TextNode (the same way `{% raw %}` does it) but quickly discovered that
we do depend on it being parsed for the StaticStylesheetAndJavascriptTags
check.

So instead, I just added a helper that both UndefinedObject and
UnusedAssign use to skip visiting of nodes that are parent'ed by a
"raw-like" tag

Fixes #685